third_party/blink/web_tests/http/tests/loading/preload-ignore-invalid-base-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
PASS internals.isPreloaded("resources/fail.js") is false
PASS window.fail is false
PASS successfullyParsed is true

TEST COMPLETE
